fix clock hand animation bug, minor changes
changed animation from css to svg. each hand as an animatetransform. setting animatetransform from and to values to degrees and degress +360 respectively. the problem before was it was animating between the current degrees and 360.

added support for tick second hand in this mode (default on, not currently configurable). calcMode is set to discrete, and the values are set with increments of 6 from 0 to 360, not counting 360, starting at the current seconds. for precision with the real digital time, this is called after waiting for 1000 minus current milliseconds, so it starts on an almost even second

defaulting to digital time with seconds for bug testing. setting digital time with new function that calls getDigitalTime on interval.

fixed AM PM bug

added options to set pin stroke and stroke width in theme.js

added --verbose to npm run dev, so the terminal shows when new stlyes are processed