fix presentation : )
presentation request param is path to receiver, previously route to index.html was a 404

currently, the clock is cast, but changes are not reflected. must use the presentation api to send and receive information so that when the user changes something, it is reflected in the receiver. or we could "cheat" and just share screen, but presentation opens the doors to a lot more. the presentation can be fullscreen with no menu or buttons, and still shown while the pc does other things in the foreground