remember tab and accordion in settings modal
modals are now always in dom, just hidden rather than not displayed. this means tabs and accordions are remembered automatically since they are always there, but this also means that we must have contents of modals (such as settings) ready when the page loads, so the navigator may have not detected the timezone etc. so we must use `?.` and default values such as 'Etc/GMT'