fix for minor onload lang bug
in theory, the header lang and the intl lang could be different. the previous code loaded the lang from the header, but set the default setting if null to the intl lang. we are now setting the default setting to the header as well, so that it is consistent (see layout and hooks files)