clean code and build output
build complained about the following styles which were fixed:

nav button.active
.dark nav button.active
.dark nav

.dark .toggle-checkbox
.dark .toggle-checkbox:checked
.dark .toggle-checkbox+.toggle-bg
.dark .toggle-checkbox:checked+.toggle-bg

note that `:global(.dark)` is necessary in Toggle.svelte but not in AccordionPanel.svelte for some reason. `@apply` with `dark:` only works in that file, and styles are not applied otherwise.

tailwind does contain a `checked:` option but we cannot check if the parent is checked, so these styles cannot all be moved inline