bug fix for mutating themes
previously, changing an attribute of the current theme before changing themes introduced a bug where the theme itself would be mutated. this appeared despite seemingly deep cloning before creating settings and using themes in many places. however, if the system is in dark mode, the theme is set to the default dark theme, and this is not a deep clone, so it is linked to the object for the default dark theme in memory and mutating the current theme mutates the default dark theme. this bug was very specific to reproduce and thus very hard to find in the first place, but has been present in production for a while