Simplify setting select component
remove font family label, should be handled by label mapper. two font family selectors in settings were changed. clock settings one doesn't have system default / empty string (because it's numerals in svg) so don't need to fix.

also, there is a bug where changing heading font family to sys default changes it to the body font if the body font isn't sys default. this was in before and after this change and the previous, and is live in prod. will add to todo