Show timezone input if autocomaticTimezone is off
Previously, the input would be disabled before having the value updated, so if the user set their timezone to new york, then turned on automatic which set it to los angeles, the input would display new york and above it would display los angeles, which could be confusing. I tried changing `bind:checked` to just `checked` and then in the `on:change` after the if statement setting locale.timezone, setting automaticTimezone to `e.target.checked`, but it still was not working.

Now, we just conditionally show the input. If it's automatic, we still want to display the timezone to the user, so we display it as text, and replace underscores with spaces